Get setProductionDomain and isDevelopment working
Fixed a bug in setProductionDomain. Setting `href` property on an `a`
tag requires a protocol to be specified or else it's assumed, when using
`hostname` that it's a relative path. Prepending the `domain` with `//`
to provide shorthand method of indicating we want to use whatever the
current protocol is.

Also fixed logic in isDevelopment.